Django Project Added
Added the project folder for the user interface of QA system.
Includes the barebones views to render the home page and the results page.
Includes a base template and a set of related pages that extend the base.

TODO: Lots of design changes. AJAX Calls to be added. Code to be cleaned up.